fix channel hash bug in query string
znc-push is kind of dumb and puts a literal hash in the query string,
which is bogus because it identifies a URI fragment, which trips up
servant.

znc-push also sends everything you put in the query string into the
request body, so now we can configure the message_uri option like:

    /activity?message={message}&network={network}&channel={context}&sender={nick}

To get a request body with all these parameters. This will decode
directly into an 'Activity'.